[Silver III] Title: 주유소, Time: 1120 ms, Memory: 191048 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -24,7 +24,7 @@ public class Main {
 		
 		for(int i=0; i<N-1; i++) {
 			min_price = Math.min(min_price, price[i]);
-			BigInteger temp = BigInteger.valueOf(min_price*distance[i]);
+			BigInteger temp = BigInteger.valueOf((long)min_price*distance[i]);
 			money = money.add(temp);
 		}
